Implemented two separate vue instance to mount the button and bopis modal and shared the data between the instance(#2vd9m35)
