+++ src/bopisPDP.ts
@@ -1,4 +1,8 @@
 import { createApp } from 'vue'
 import Bopis from './Bopis.vue'
+import BopisModal from './BopisModal.vue'
 
-createApp(Bopis).mount('#hc-bopis-app');
+const bopisModalInstance = createApp(BopisModal).mount('#hc-bopis-app');
+const bopisButtonInstance = createApp(Bopis).mount('#hc-bopis-button')
+
+export { bopisButtonInstance, bopisModalInstance }
